Set up draft art for image switching, rerolled selections to prevent same flavor from being picked twice in a row
In addition to setting up image switching, I also revised HTML so images are placed as imgs inside divs. This enables the alt text to be updated dynamically when the cats animate. The previous setup updated the background image of the div, which could accomodate aria labels, but alt/title is more typical for images.